BB-21555: Filters "is not any of" and 'is any of" round user-entered value (#33423)

+++ src/Oro/Bundle/FilterBundle/Resources/public/js/filter/number-filter.js
@@ -130,7 +130,7 @@ define(function(require) {
                 _.map(
                     value.toString().split(this.arraySeparator),
                     function(number) {
-                        return parseInt(number);
+                        return parseFloat(number);
                     }
                 ),
                 function(number) {
